+++ green_city/data/schema/error_event_schema.py
