+++ green_city/fixtures/create_event_fixtures.py
@@ -16,13 +16,10 @@ def create_event(auth_token):
 
     response = requests.post(url, headers={"Authorization": auth_token}, files=files)
 
-    assert response.status_code == 201
-    assert response.headers['Content-Type'] == 'application/json'
-
     event_data = response.json()
     event_id = event_data.get("id")
     yield event_data
 
     delete_url = f"{API_BASE_URL_8085}{ENDPOINTS['delete_events'].format(event_id)}"
     del_response = requests.delete(delete_url, headers={"Authorization": auth_token})
-    assert del_response.status_code == 200, "Failed to delete event"
+
