+++ green_city/tests/api/post_event_test.py
@@ -2,22 +2,20 @@ import json
 import pytest
 import requests
 from jsonschema import validate
-from green_city.src.config import API_BASE_URL_8085, ENDPOINTS
-from green_city.tests.data.fixture_dto.create_event_dto_request import create_event_dto_request
-from green_city.tests.data.schema.error_event_schema import error_event_schemas
-from green_city.tests.data.schema.event_schema import EVENT_SCHEMA
-import green_city.src.util.logging_config
-import logging
-
-from green_city.tests.data.test_data.post_event_error_message import post_event_test_data
-
-logger = logging.getLogger(__name__)
+from green_city.config.config import API_BASE_URL_8085, ENDPOINTS
+from green_city.data.schema.error_event_schema import error_event_schemas
+from green_city.data.schema.event_schema import EVENT_SCHEMA
+from green_city.fixtures.create_event_fixtures import create_event
+from green_city.fixtures.login_fixture import auth_token
+from green_city.data.test_data.post_event_error_message import post_event_test_data
+from green_city.data.fixture_dto.create_event_dto_request import create_event_dto_request
 
 def test_create_event_returns_201(create_event):
     event_data = create_event
 
-    validate(instance=event_data, schema=EVENT_SCHEMA)
-
+    assert validate(instance=event_data, schema=EVENT_SCHEMA) is None, (
+        "Response data does not match schema"
+    )
 
 def test_create_event_unauthorized_returns_401():
 
@@ -32,16 +30,11 @@ def test_create_event_unauthorized_returns_401():
 
     response = requests.post(url, files=files)
 
-    logger.info(f"Status code: {response.status_code}")
-    logger.info(f"Response body: {response.text}")
-
     assert response.status_code == 401, (
         f"Expected status code 401, got {response.status_code}"
     )
 
     response_data = response.json()
-
-    logger.debug(f"Parsed response JSON: {json.dumps(response_data, indent=2)}")
 
     assert response_data['status'] == 401, "Status field should be 401"
     assert response_data['error'] == "Unauthorized", "Error field should be 'Unauthorized'"
@@ -69,18 +62,15 @@ def test_create_event_returns_400(test_case_key, auth_token):
 
     response = requests.post(url, files=files, headers=headers)
 
-    logger.info(f"Status code: {response.status_code}")
-    logger.info(f"Response body: {response.text}")
-
     assert response.status_code == 400, f"Expected 400 for '{test_case_key}', got {response.status_code}"
 
     response_json = response.json()
     if isinstance(response_json, dict):
         response_json = [response_json]
 
-    logger.debug(f"Parsed response JSON: {json.dumps(response_json, indent=2)}")
-
-    validate(instance=response_json, schema=error_event_schemas["400_validation_errors"])
+    assert validate(instance=response_json, schema=error_event_schemas["400_validation_errors"]) is None, (
+        "Response data does not match schema"
+    )
 
     for expected_error in test_case["expected_errors"]:
         assert expected_error in response_json, (
